feat: Add various improvements and fixes
- Update gdino_teacher.py with path and dtype fixes
- Update finetune.py with alignment training changes
- Update action_heads.py for better action prediction
- Update visual teacher configs and scripts
- Add data processing scripts
- Add run.sh training script

+++ prismatic/models/action_heads.py
@@ -50,7 +50,7 @@ class L1RegressionActionHead(nn.Module):
         batch_size = actions_hidden_states.shape[0]
         device = actions_hidden_states.device
 
-        proprio = proprio.reshape(batch_size, -1).to(torch.bfloat16)  # (bsz, proprio_dim)
+        proprio = proprio.reshape(batch_size, -1).to(actions_hidden_states.dtype)  # (bsz, proprio_dim)
         proprio_features = proprio_projector(proprio)  # (bsz, llm_dim)
         proprio_features = proprio_features.unsqueeze(dim=1)  # (bsz, 1, llm_dim)
